refactor: Split massive chat.go into focused modules
Break down the 654-line internal/tui/chat.go into smaller,
focused modules within internal/tui/chat/ package:

- model.go: ChatModel struct and constructor
- update.go: Update method logic and message handling
- view.go: View method and rendering logic
- handlers.go: Request/response handling
- selection.go: Model/provider selection functionality
- history.go: Input history navigation
- suggestions.go: Command suggestion logic
- commands.go: Command processing and handlers
- types.go: Shared message types

Also updates cmd/rigel/main.go to use chat.NewModel() directly
instead of going through tui package.
