--- internal/tui/chat.go
@@ -1,654 +1,0 @@
-package tui
-
-import (
-	"context"
-	"fmt"
-	"strings"
-
-	"github.com/charmbracelet/bubbles/spinner"
-	"github.com/charmbracelet/bubbles/textarea"
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/mizzy/rigel/internal/config"
-	"github.com/mizzy/rigel/internal/history"
-	"github.com/mizzy/rigel/internal/llm"
-	"github.com/mizzy/rigel/internal/tui/render"
-)
-
-// ChatModel represents the main chat interface
-type ChatModel struct {
-	provider llm.Provider
-	config   *config.Config
-	input    textarea.Model
-	spinner  spinner.Model
-
-	history            []Exchange
-	inputHistory       []string
-	historyIndex       int
-	currentInput       string
-	thinking           bool
-	currentPrompt      string
-	err                error
-	quitting           bool
-	suggestions        []string
-	selectedSuggestion int
-	showSuggestions    bool
-	ctrlCPressed       bool
-	infoMessage        string
-	historyManager     *history.Manager // Add history manager
-
-	// Model selection mode
-	modelSelectionMode bool
-	availableModels    []llm.Model
-	filteredModels     []llm.Model
-	selectedModelIndex int
-	modelFilter        string
-
-	// Provider selection mode
-	providerSelectionMode bool
-	availableProviders    []string
-	selectedProviderIndex int
-}
-
-// Exchange represents a single chat exchange
-type Exchange struct {
-	Prompt   string
-	Response string
-}
-
-// NewChatModel creates a new chat model instance
-func NewChatModel(provider llm.Provider, cfg *config.Config) *ChatModel {
-	ta := textarea.New()
-	ta.Placeholder = "Type a message or / for commands (Alt+Enter for new line)"
-	ta.Focus()
-	ta.CharLimit = 5000
-	ta.SetWidth(100)
-	ta.SetHeight(3)
-	ta.ShowLineNumbers = false
-	ta.Prompt = ""                // Remove the default prompt
-	ta.EndOfBufferCharacter = ' ' // Use space instead of default character
-	ta.KeyMap.InsertNewline.SetKeys("alt+enter", "ctrl+j")
-
-	// Remove borders and styling
-	noBorder := lipgloss.NewStyle().
-		Border(lipgloss.NormalBorder(), false).
-		BorderForeground(lipgloss.NoColor{}).
-		Padding(0).
-		Margin(0)
-
-	ta.FocusedStyle.Base = noBorder
-	ta.FocusedStyle.CursorLine = lipgloss.NewStyle()
-	ta.FocusedStyle.Placeholder = lipgloss.NewStyle().Foreground(lipgloss.Color("60")) // Dim blue
-	ta.FocusedStyle.Prompt = lipgloss.NewStyle()
-	ta.FocusedStyle.Text = lipgloss.NewStyle()
-	ta.FocusedStyle.EndOfBuffer = lipgloss.NewStyle()
-
-	ta.BlurredStyle.Base = noBorder
-	ta.BlurredStyle.CursorLine = lipgloss.NewStyle()
-	ta.BlurredStyle.Placeholder = lipgloss.NewStyle().Foreground(lipgloss.Color("60")) // Dim blue
-	ta.BlurredStyle.Prompt = lipgloss.NewStyle()
-	ta.BlurredStyle.Text = lipgloss.NewStyle()
-	ta.BlurredStyle.EndOfBuffer = lipgloss.NewStyle()
-
-	s := spinner.New()
-	s.Spinner = spinner.Dot
-	s.Style = lipgloss.NewStyle().Foreground(lipgloss.Color("87")) // Match Rigel blue
-
-	// Initialize history manager
-	histManager, err := history.NewManager()
-	if err != nil {
-		// If we can't create history manager, continue without it
-		histManager = nil
-	} else {
-		// Load existing history
-		_ = histManager.Load()
-	}
-
-	m := &ChatModel{
-		provider:       provider,
-		config:         cfg,
-		input:          ta,
-		spinner:        s,
-		history:        []Exchange{},
-		inputHistory:   []string{},
-		historyIndex:   -1,
-		historyManager: histManager,
-	}
-
-	// Load input history from manager if available
-	if histManager != nil {
-		m.inputHistory = histManager.GetCommands()
-	}
-
-	return m
-}
-
-// Init initializes the chat model
-func (m ChatModel) Init() tea.Cmd {
-	return tea.Batch(
-		textarea.Blink,
-		m.spinner.Tick,
-	)
-}
-
-// Update handles messages and updates the model
-func (m ChatModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	var cmd tea.Cmd
-	var cmds []tea.Cmd
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		// Handle provider selection mode
-		if m.providerSelectionMode {
-			return m.handleProviderSelectionKey(msg)
-		}
-
-		// Handle model selection mode
-		if m.modelSelectionMode {
-			return m.handleModelSelectionKey(msg)
-		}
-
-		// Handle special keys first
-		switch msg.Type {
-		case tea.KeyCtrlC:
-			if m.ctrlCPressed {
-				m.quitting = true
-				return m, tea.Quit
-			}
-			m.ctrlCPressed = true
-			m.infoMessage = "Press Ctrl+C again to exit"
-			// Reset Ctrl+C flag after any other key
-			return m, nil
-
-		case tea.KeyCtrlD:
-			// Reset Ctrl+C flag on any other key
-			m.ctrlCPressed = false
-			m.infoMessage = ""
-			if !m.thinking && m.input.Value() == "" {
-				m.quitting = true
-				return m, tea.Quit
-			}
-		default:
-			// Reset Ctrl+C flag on any other key
-			m.ctrlCPressed = false
-			m.infoMessage = ""
-		}
-
-		// Handle Tab key for completion
-		if msg.String() == "tab" && !m.thinking && m.showSuggestions {
-			m.completeSuggestion()
-			m.ctrlCPressed = false // Reset Ctrl+C flag
-			m.infoMessage = ""
-			return m, nil
-		}
-
-		// Handle arrow keys for suggestion navigation or history navigation
-		if !m.thinking {
-			switch msg.String() {
-			case "up":
-				if m.showSuggestions {
-					if m.selectedSuggestion > 0 {
-						m.selectedSuggestion--
-					}
-				} else {
-					m.navigateHistory(-1)
-				}
-				m.ctrlCPressed = false // Reset Ctrl+C flag
-				m.infoMessage = ""
-				return m, nil
-			case "down":
-				if m.showSuggestions {
-					if m.selectedSuggestion < len(m.suggestions)-1 {
-						m.selectedSuggestion++
-					}
-				} else {
-					m.navigateHistory(1)
-				}
-				m.ctrlCPressed = false // Reset Ctrl+C flag
-				m.infoMessage = ""
-				return m, nil
-			}
-		}
-
-		// Check for Enter key specifically (not Alt+Enter)
-		if msg.String() == "enter" && !m.thinking {
-			m.ctrlCPressed = false // Reset Ctrl+C flag
-			m.infoMessage = ""
-			// If suggestions are shown and one is selected, complete and execute it
-			if m.showSuggestions {
-				m.completeSuggestion()
-				// After completing suggestion, check if it's a command and execute it
-				if strings.HasPrefix(m.input.Value(), "/") {
-					// Treat it as if user pressed Enter with the command
-					m.currentPrompt = m.input.Value()
-
-					// Save to input history
-					m.inputHistory = append(m.inputHistory, m.currentPrompt)
-					m.historyIndex = -1
-					m.currentInput = ""
-
-					// Save to persistent history
-					if m.historyManager != nil {
-						_ = m.historyManager.Add(m.currentPrompt)
-					}
-
-					m.input.SetValue("")
-					m.thinking = true
-					m.err = nil
-					m.showSuggestions = false
-
-					// Handle the command
-					trimmedPrompt := strings.TrimSpace(m.currentPrompt)
-					cmd := m.handleCommand(trimmedPrompt)
-					if cmd != nil {
-						return m, tea.Batch(cmd, m.spinner.Tick)
-					}
-				}
-				return m, nil
-			}
-
-			if strings.TrimSpace(m.input.Value()) != "" {
-				m.currentPrompt = m.input.Value()
-
-				// Save to input history
-				m.inputHistory = append(m.inputHistory, m.currentPrompt)
-				m.historyIndex = -1
-				m.currentInput = ""
-
-				// Save to persistent history
-				if m.historyManager != nil {
-					_ = m.historyManager.Add(m.currentPrompt)
-				}
-
-				m.input.SetValue("")
-				m.thinking = true
-				m.err = nil
-				m.showSuggestions = false
-
-				// Handle commands
-				trimmedPrompt := strings.TrimSpace(m.currentPrompt)
-				cmd := m.handleCommand(trimmedPrompt)
-				if cmd != nil {
-					return m, tea.Batch(cmd, m.spinner.Tick)
-				}
-			}
-			return m, nil
-		}
-
-		// Pass all other keys (including alt+enter and ctrl+j) to textarea
-		if !m.thinking {
-			oldValue := m.input.Value()
-			m.input, cmd = m.input.Update(msg)
-
-			// Update suggestions if input changed
-			if oldValue != m.input.Value() {
-				m.updateSuggestions()
-				m.ctrlCPressed = false // Reset Ctrl+C flag when typing
-				m.infoMessage = ""
-
-				// Reset history navigation if user types
-				if m.historyIndex != -1 {
-					m.historyIndex = -1
-					m.currentInput = m.input.Value()
-				}
-			}
-
-			return m, cmd
-		}
-
-	case providerSelectorMsg:
-		if msg.err != nil {
-			return m, func() tea.Msg {
-				return aiResponse{err: msg.err}
-			}
-		}
-
-		m.providerSelectionMode = true
-		m.availableProviders = msg.providers
-		m.selectedProviderIndex = 0
-
-		// Find current provider index
-		for i, p := range msg.providers {
-			if p == msg.currentProvider {
-				m.selectedProviderIndex = i
-				break
-			}
-		}
-
-		return m, nil
-
-	case modelSelectorMsg:
-		if msg.err != nil {
-			return m, func() tea.Msg {
-				return aiResponse{err: msg.err}
-			}
-		}
-
-		m.modelSelectionMode = true
-		m.availableModels = msg.models
-		m.filteredModels = msg.models
-		m.selectedModelIndex = 0
-		m.modelFilter = ""
-		m.input.SetValue("")
-		m.input.Placeholder = "Type to filter models, Enter to select, Esc to cancel"
-
-		return m, nil
-
-	case providerSwitchResponse:
-		m.provider = msg.provider
-		m.thinking = false
-		m.history = append(m.history, Exchange{
-			Prompt:   m.currentPrompt,
-			Response: fmt.Sprintf("Switched to provider: %s\nCurrent model: %s", msg.providerName, m.provider.GetCurrentModel()),
-		})
-		m.currentPrompt = ""
-		return m, nil
-
-	case aiResponse:
-		m.thinking = false
-
-		if msg.err != nil {
-			m.err = msg.err
-		} else {
-			m.history = append(m.history, Exchange{
-				Prompt:   m.currentPrompt,
-				Response: msg.content,
-			})
-			m.currentPrompt = ""
-		}
-		return m, nil
-
-	case spinner.TickMsg:
-		if m.thinking {
-			m.spinner, cmd = m.spinner.Update(msg)
-			return m, cmd
-		}
-	}
-
-	if !m.thinking {
-		m.input, cmd = m.input.Update(msg)
-		cmds = append(cmds, cmd)
-	}
-
-	return m, tea.Batch(cmds...)
-}
-
-func (m *ChatModel) handleProviderSelectionKey(msg tea.KeyMsg) (*ChatModel, tea.Cmd) {
-	switch msg.Type {
-	case tea.KeyEsc:
-		m.exitProviderSelection()
-		return m, nil
-
-	case tea.KeyEnter:
-		if m.selectedProviderIndex < len(m.availableProviders) {
-			selectedProvider := m.availableProviders[m.selectedProviderIndex]
-			m.exitProviderSelection()
-			return m, m.switchProvider(selectedProvider)
-		}
-		return m, nil
-
-	case tea.KeyUp:
-		if m.selectedProviderIndex > 0 {
-			m.selectedProviderIndex--
-		}
-		return m, nil
-
-	case tea.KeyDown:
-		if m.selectedProviderIndex < len(m.availableProviders)-1 {
-			m.selectedProviderIndex++
-		}
-		return m, nil
-
-	default:
-		return m, nil
-	}
-}
-
-func (m *ChatModel) handleModelSelectionKey(msg tea.KeyMsg) (*ChatModel, tea.Cmd) {
-	switch msg.Type {
-	case tea.KeyEsc:
-		m.exitModelSelection()
-		return m, nil
-
-	case tea.KeyEnter:
-		if len(m.filteredModels) > 0 && m.selectedModelIndex < len(m.filteredModels) {
-			selectedModel := m.filteredModels[m.selectedModelIndex]
-			m.exitModelSelection()
-			return m, m.switchModel(selectedModel.Name)
-		}
-		return m, nil
-
-	case tea.KeyUp:
-		if m.selectedModelIndex > 0 {
-			m.selectedModelIndex--
-		}
-		return m, nil
-
-	case tea.KeyDown:
-		if m.selectedModelIndex < len(m.filteredModels)-1 {
-			m.selectedModelIndex++
-		}
-		return m, nil
-
-	default:
-		var cmd tea.Cmd
-		m.input, cmd = m.input.Update(msg)
-
-		newFilter := m.input.Value()
-		if newFilter != m.modelFilter {
-			m.modelFilter = newFilter
-			m.filterModels()
-			m.selectedModelIndex = 0
-		}
-
-		return m, cmd
-	}
-}
-
-func (m *ChatModel) exitProviderSelection() {
-	m.providerSelectionMode = false
-	m.availableProviders = nil
-	m.selectedProviderIndex = 0
-	m.thinking = false
-}
-
-func (m *ChatModel) exitModelSelection() {
-	m.modelSelectionMode = false
-	m.input.SetValue("")
-	m.input.Placeholder = "Type a message or / for commands (Alt+Enter for new line)"
-	m.modelFilter = ""
-	m.filteredModels = nil
-	m.availableModels = nil
-	m.selectedModelIndex = 0
-	m.thinking = false
-}
-
-func (m *ChatModel) filterModels() {
-	if m.modelFilter == "" {
-		m.filteredModels = m.availableModels
-		return
-	}
-
-	filter := strings.ToLower(m.modelFilter)
-	m.filteredModels = nil
-
-	for _, model := range m.availableModels {
-		if strings.Contains(strings.ToLower(model.Name), filter) {
-			m.filteredModels = append(m.filteredModels, model)
-		}
-	}
-}
-
-func (m *ChatModel) switchProvider(providerName string) tea.Cmd {
-	return func() tea.Msg {
-		// Update config
-		if m.config != nil {
-			m.config.Provider = providerName
-		}
-
-		// Create new provider
-		newProvider, err := llm.NewProvider(m.config)
-		if err != nil {
-			return aiResponse{err: fmt.Errorf("failed to switch provider: %w", err)}
-		}
-
-		return providerSwitchResponse{
-			provider:     newProvider,
-			providerName: providerName,
-		}
-	}
-}
-
-func (m *ChatModel) switchModel(modelName string) tea.Cmd {
-	// Actually switch the model
-	m.provider.SetModel(modelName)
-
-	return func() tea.Msg {
-		return aiResponse{
-			content: fmt.Sprintf("Switched to model: %s", modelName),
-		}
-	}
-}
-
-// View renders the chat interface
-func (m ChatModel) View() string {
-	if m.quitting {
-		return ""
-	}
-
-	var s strings.Builder
-
-	// Render chat history using extracted render function
-	renderHistory := make([]render.Exchange, len(m.history))
-	for i, ex := range m.history {
-		renderHistory[i] = render.Exchange{
-			Prompt:   ex.Prompt,
-			Response: ex.Response,
-		}
-	}
-	s.WriteString(render.ChatHistory(renderHistory))
-
-	// Display provider selection interface if in provider selection mode
-	if m.providerSelectionMode {
-		return render.ProviderSelector(m.availableProviders, m.selectedProviderIndex)
-	}
-
-	// Display model selection interface if in model selection mode
-	if m.modelSelectionMode {
-		return render.ModelSelector(m.filteredModels, m.selectedModelIndex, m.modelFilter)
-	}
-
-	// Display thinking state
-	if m.thinking {
-		s.WriteString(render.ThinkingState(m.currentPrompt, m.spinner.View()))
-	}
-
-	// Display input prompt and suggestions
-	if !m.thinking {
-		s.WriteString(render.InputPrompt(m.input.View()))
-
-		// Display command suggestions using render function
-		if m.showSuggestions && len(m.suggestions) > 0 {
-			// Convert commands to render.Command format
-			renderCommands := make([]render.Command, len(availableCommands))
-			for i, cmd := range availableCommands {
-				renderCommands[i] = render.Command{
-					Command:     cmd.command,
-					Description: cmd.description,
-				}
-			}
-			s.WriteString(render.CommandSuggestions(m.suggestions, m.selectedSuggestion, renderCommands))
-		}
-	}
-
-	// Display messages using render functions
-	s.WriteString(render.InfoMessage(m.infoMessage))
-	s.WriteString(render.ErrorMessage(m.err))
-
-	return s.String()
-}
-
-// requestResponse sends a request to the LLM provider with conversation history
-func (m *ChatModel) requestResponse(prompt string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-
-		// Build message history from chat exchanges
-		messages := make([]llm.Message, 0, len(m.history)*2+1)
-
-		// Add previous exchanges to maintain context
-		for _, exchange := range m.history {
-			messages = append(messages, llm.Message{
-				Role:    "user",
-				Content: exchange.Prompt,
-			})
-			messages = append(messages, llm.Message{
-				Role:    "assistant",
-				Content: exchange.Response,
-			})
-		}
-
-		// Add current prompt
-		messages = append(messages, llm.Message{
-			Role:    "user",
-			Content: prompt,
-		})
-
-		// Use GenerateWithHistory to send full conversation context
-		response, err := m.provider.GenerateWithHistory(ctx, messages, llm.GenerateOptions{})
-		if err != nil {
-			return aiResponse{err: err}
-		}
-		return aiResponse{content: strings.TrimSpace(response)}
-	}
-}
-
-// aiResponse represents a response from the AI
-type aiResponse struct {
-	content string
-	err     error
-}
-
-// providerSwitchResponse represents a provider switch response
-type providerSwitchResponse struct {
-	provider     llm.Provider
-	providerName string
-}
-
-// navigateHistory navigates through input history
-func (m *ChatModel) navigateHistory(direction int) {
-	if len(m.inputHistory) == 0 {
-		return
-	}
-
-	// Save current input if we're starting to navigate history
-	if m.historyIndex == -1 {
-		m.currentInput = m.input.Value()
-	}
-
-	if direction < 0 {
-		// Going up (backward) in history
-		if m.historyIndex == -1 {
-			// Start from the most recent item
-			m.historyIndex = 0
-		} else if m.historyIndex < len(m.inputHistory)-1 {
-			m.historyIndex++
-		}
-
-		if m.historyIndex < len(m.inputHistory) {
-			historyPos := len(m.inputHistory) - 1 - m.historyIndex
-			m.input.SetValue(m.inputHistory[historyPos])
-		}
-	} else {
-		// Going down (forward) in history
-		if m.historyIndex > 0 {
-			m.historyIndex--
-			historyPos := len(m.inputHistory) - 1 - m.historyIndex
-			m.input.SetValue(m.inputHistory[historyPos])
-		} else if m.historyIndex == 0 {
-			// Return to current input
-			m.historyIndex = -1
-			m.input.SetValue(m.currentInput)
-		}
-	}
-}
