+++ internal/tui/chat/types.go
@@ -1,0 +1,29 @@
+package chat
+
+import "github.com/mizzy/rigel/internal/llm"
+
+// aiResponse represents a response from the AI
+type aiResponse struct {
+	content string
+	err     error
+}
+
+// providerSwitchResponse represents a provider switch response
+type providerSwitchResponse struct {
+	provider     llm.Provider
+	providerName string
+}
+
+// modelSelectorMsg is sent when model selection is requested
+type modelSelectorMsg struct {
+	models       []llm.Model
+	currentModel string
+	err          error
+}
+
+// providerSelectorMsg is sent when provider selection is requested
+type providerSelectorMsg struct {
+	providers       []string
+	currentProvider string
+	err             error
+}
